% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state Plot map of accidents in selected state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing state number, class: integer}

\item{year}{An integer representing selected year, class: integer}
}
\value{
Returns "no accidents to plot" if there were no accidents. Otherwise returns a map containing the accident locations for selected data.
}
\description{
This function exports the map of accidents during the given year. If no accients exist, function returns "no accidents to plot".
}
\note{
Stops execution if no state exists with the given state.num.
}
\examples{
fars_map_state(10, 2010)

}
