% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read data file reader}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Filename in csv. format., class: character}
}
\value{
This function returns a tibble dataframe of the loaded data with dplyr package.
}
\description{
This is a function that reads in a file with FARS data and informs user if the file does not exist as expected. Function can be used to
read a desired filename, but usually this function is used together with a year by function fars_read_years()
}
\note{
Stops execution if file does not exis e.g. no data for the given year exists.
}
\examples{
fars_read(file.csv)

}
