% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years reats a summary of the data with selected years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{selected years as vector, class: integer}
}
\value{
a summary table of the data
}
\description{
This function creates summary data of accidents based on selected years as form of a table.
}
\examples{
fars_summarize_years(c(2010, 2011))

}
