% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename filename generator helper function}
\usage{
make_filename(year)
}
\arguments{
\item{year, }{class: integer}
}
\value{
generated filename with year to be read, class: character
}
\description{
make_filename generates a filename of format "accident_%d.csv.bz2NNNN", where NNNN=year that is read based on user input of the year parameter. Used as helper -function.
}
\examples{
make_filename(2010)
}
